Happy ... New Year?
Yeah, we haven't updated this repo for nearly half a year but we are actually still maintaining it. We are now using them in our commercial products, and constantly improving Holakit, including the long-left-open PR for Holakit 2.x. We just need more time to extract a bunch of new components and improvements from our delivered products, and write documents for them. We are still there, yes, Team ¡Hola! is also still there too. In 2019 I will be 20, (Oh, poor naive youth!) and we are still figuring out a way to properly address our mission. This commit is another beginning, and we aren't defeated by life yet.